fix surefire error
