+++ sdk/cosmos/azure-cosmos-tests/src/test/java/com/azure/cosmos/implementation/ConfigsTests.java
@@ -69,14 +69,4 @@ public class ConfigsTests {
             System.clearProperty("COSMOS.METRICS_CONFIG");
         }
     }
-
-
-    @Test(groups = { "unit" })
-    public void allowUnquotedControlChars() {
-        assertThat(Configs.shouldAllowUnquotedControlChars()).isTrue();
-
-        System.setProperty("COSMOS.ALLOW_UNQUOTED_CONTROL_CHARS", "false");
-        assertThat(Configs.shouldAllowUnquotedControlChars()).isFalse();
-        System.clearProperty("COSMOS.ALLOW_UNQUOTED_CONTROL_CHARS");
-    }
 }
