+++ sdk/cosmos/azure-cosmos/src/main/java/com/azure/cosmos/implementation/Utils.java
@@ -51,8 +51,10 @@ import java.util.concurrent.atomic.AtomicReference;
 import java.util.function.Consumer;
 import java.util.regex.Pattern;
 
+import static com.azure.cosmos.implementation.guava25.base.MoreObjects.firstNonNull;
 import static com.azure.cosmos.implementation.guava25.base.Preconditions.checkArgument;
 import static com.azure.cosmos.implementation.guava25.base.Preconditions.checkNotNull;
+import static com.azure.cosmos.implementation.guava25.base.Strings.emptyToNull;
 
 /**
  * While this class is public, but it is not part of our published public APIs.
@@ -60,6 +62,11 @@ import static com.azure.cosmos.implementation.guava25.base.Preconditions.checkNo
  */
 public class Utils {
     private final static Logger logger = LoggerFactory.getLogger(Utils.class);
+
+    // Flag to indicate whether enable JsonParser.Feature.ALLOW_UNQUOTED_CONTROL_CHARS
+    // Keep the config here not Configs to break the circular reference
+    private static final boolean DEFAULT_ALLOW_UNQUOTED_CONTROL_CHARS = true;
+    private static final String ALLOW_UNQUOTED_CONTROL_CHARS = "COSMOS.ALLOW_UNQUOTED_CONTROL_CHARS";
 
     public static final Class<?> byteArrayClass = new byte[0].getClass();
 
@@ -115,7 +122,7 @@ public class Utils {
             objectMapper.configure(JsonParser.Feature.STRICT_DUPLICATE_DETECTION, true);
         }
         objectMapper.configure(DeserializationFeature.ACCEPT_FLOAT_AS_INT, false);
-        if (Configs.shouldAllowUnquotedControlChars()) {
+        if (shouldAllowUnquotedControlChars()) {
             objectMapper.configure(JsonParser.Feature.ALLOW_UNQUOTED_CONTROL_CHARS, true);
         }
 
@@ -764,5 +771,17 @@ public class Utils {
             throw new IllegalArgumentException("MaxIntegratedCacheStaleness duration cannot be negative");
         }
         return maxIntegratedCacheStaleness.toMillis();
+    }
+
+    public static boolean shouldAllowUnquotedControlChars() {
+
+        String shouldAllowUnquotedControlCharsConfig =
+            System.getProperty(
+                ALLOW_UNQUOTED_CONTROL_CHARS,
+                firstNonNull(
+                    emptyToNull(System.getenv().get(ALLOW_UNQUOTED_CONTROL_CHARS)),
+                    String.valueOf(DEFAULT_ALLOW_UNQUOTED_CONTROL_CHARS)));
+
+        return Boolean.parseBoolean(shouldAllowUnquotedControlCharsConfig);
     }
 }
